Refactored Code: Put classes in model package
