+++ .gitignore
@@ -159,4 +159,4 @@ tags
 
 # End of https://www.toptal.com/developers/gitignore/api/python,vim
 
-
+.vscode
