+++ model/User.py
@@ -1,0 +1,8 @@
+class User():
+    def __init__(self, id, username, password):
+        self.id = id
+        self.username = username
+        self.password = password
+
+    def __str__(self):
+        return "User(id='%s')" % self.id
